+++ EssentialFeed/EssentialFeediOSTests/Helpers/FeedViewControllerTests+LoaderSpy.swift
@@ -1,0 +1,65 @@
+//
+//  FeedViewControllerTests+LoaderSpy.swift
+//  EssentialFeediOSTests
+//
+//  Created by nicaho on 2024/8/6.
+//
+
+import Foundation
+import EssentialFeed
+import EssentialFeediOS
+
+class LoaderSpy: FeedLoader, FeedImageDataLoader {
+    var loadCallCount: Int {
+        feedRequest.count
+    }
+    
+    private var feedRequest = [(FeedLoader.Result) -> Void]()
+    
+    func load(completion: @escaping (FeedLoader.Result) -> Void) {
+        feedRequest.append(completion)
+    }
+    
+    func completeFeedLoading(with feed: [FeedImage] = [], at index: Int = 0) {
+        feedRequest[index](.success(feed))
+    }
+    
+    func completeFeedLoadingWithError(at index: Int = 0) {
+        let error = NSError(domain: "an error", code: 0)
+        feedRequest[index](.failure(error))
+    }
+    
+    // MARK: - FeedImageDataLoader
+    
+    private struct TaskSpy: FeedImageDataLoaderTask {
+        let cancelCallback: () -> Void
+        func cancel() {
+            cancelCallback()
+        }
+    }
+    
+    var loadedImageURLs: [URL] {
+        imageRequests.map { $0.url }
+    }
+    
+    private var imageRequests = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
+    private(set) var cancelledImageURLs = [URL]()
+    
+    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        self.imageRequests.append((url, completion))
+        return TaskSpy { [weak self] in
+            self?.cancelledImageURLs.append(url)
+        }
+    }
+    
+    func completeImageLoading(with imageData: Data = Data(), at index: Int = 0) {
+        imageRequests[index].completion(.success(imageData))
+    }
+    
+    func completeImageLoadingWithError(at index: Int = 0) {
+        let error = NSError(domain: "an error", code: 0)
+        imageRequests[index].completion(.failure(error))
+    }
+    
+    
+}
